Move backend status to advanced settings
